Structure established

+++ src/main/java/com/hank/pizzSystem/hankPizzaSystem/entity/Info/Cheese.java
@@ -20,7 +20,7 @@ public class Cheese {
     private String cheeseName;
 
     @Column(nullable = false, unique = false)
-    private float unitPrice;
+    private int unitPrice;
  //   @Column(nullable = false, unique = true)
    // private int Stock;
